Implementing Signing in.
This commit works on implementing the ability to signin by creating the
sign in route. The route looks the user up by their email, verfies their
password is correct by using the compare password function, and will send
a helpful message back depending on wether or not the user was signed in.