Finishing up messages.
This commit implements GET and DELETE for messages. The route is
/api/users/:id/messages/:message_id and GET or DELETE request.
This commit also created a route to list out all the messages.
The route is /api/messages. The user must be logged in to use
this route. Finally, the message model was updated to use
timestamps. This should finish the backend for the project.